+++ modules/gltools/src/context/context.d.ts
@@ -9,24 +9,26 @@ export type CreateGLContextOptions = {
   webgl1?: boolean; // set to false to not create a WebGL1 context (fail if webgl2 not available)
   throwOnError?: boolean; // If set to false, return `null` if context creation fails.
   manageState?: boolean; // Set to false to disable WebGL state management instrumentation
-  break?: Array<any> // TODO: types
+  break?: Array<any>; // TODO: types
   onError?: any;
+  onContextLost?: (event: Event) => void;
+  onContextRestored?: (event: Event) => void;
   // BROWSER CONTEXT PARAMETERS
-  canvas?: HTMLCanvasElement | string | OffscreenCanvas | null // A canvas element or a canvas string id
-  debug?: boolean // Instrument context (at the expense of performance)
-  alpha?: boolean // Default render target has an alpha buffer.
-  depth?: boolean // Default render target has a depth buffer of at least 16 bits.
-  stencil?: boolean // Default render target has a stencil buffer of at least 8 bits.
-  antialias?: boolean // Boolean that indicates whether or not to perform anti-aliasing.
-  premultipliedAlpha?: boolean // Boolean that indicates that the page compositor will assume the drawing buffer contains colors with pre-multiplied alpha.
-  preserveDrawingBuffer?: boolean // Default render target buffers will not be automatically cleared and will preserve their values until cleared or overwritten
-  failIfMajorPerformanceCaveat?: boolean // Do not create if the system performance is low.
+  canvas?: HTMLCanvasElement | string | OffscreenCanvas | null; // A canvas element or a canvas string id
+  debug?: boolean; // Instrument context (at the expense of performance)
+  alpha?: boolean; // Default render target has an alpha buffer.
+  depth?: boolean; // Default render target has a depth buffer of at least 16 bits.
+  stencil?: boolean; // Default render target has a stencil buffer of at least 8 bits.
+  antialias?: boolean; // Boolean that indicates whether or not to perform anti-aliasing.
+  premultipliedAlpha?: boolean; // Boolean that indicates that the page compositor will assume the drawing buffer contains colors with pre-multiplied alpha.
+  preserveDrawingBuffer?: boolean; // Default render target buffers will not be automatically cleared and will preserve their values until cleared or overwritten
+  failIfMajorPerformanceCaveat?: boolean; // Do not create if the system performance is low.
   // HEADLESS CONTEXT PARAMETERS
   width?: number /** width are height are only used by headless gl */;
   height?: number /** width are height are only used by headless gl */;
   // WEBGL/HEADLESS CONTEXT PARAMETERS
   // Remaining options are passed through to context creator
-}
+};
 
 /*
  * Creates a context giving access to the WebGL API